requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
dacite==1.8.1
dateparser==1.2.1
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
executing==2.1.0
filelock==3.17.0
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
frozenlist==1.5.0
fsspec==2025.2.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.7
ImageHash==4.3.1
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.3.21
langchain-core==0.3.48
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langsmith==0.3.18
Levenshtein==0.27.1
llvmlite==0.42.0
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multimethod==1.11.2
mysqlclient==2.2.5
narwhals==1.32.0
nashpy==0.0.41
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.68.2
orjson==3.10.16
packaging==24.0
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.3.0
platformdirs==4.3.6
plotly==6.0.1
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytrends==4.9.2
pytz==2025.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.2.0
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3
urllib3==2.2.1
visions==0.7.6
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.0.4
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.18.3
ydata-profiling==4.8.3
zstandard==0.23.0
